+++ src/basins/fractality_of_basins.jl
@@ -1,10 +1,13 @@
 export uncertainty_exponent, basins_fractal_dimension, basins_fractal_test, basin_entropy
 
 """
+    basin_entropy(BoA::ArrayBasinsOfAttraction{Integer}, ε = size(BoA.basins, 1)÷10) -> Sb, Sbb
     basin_entropy(basins::Array{Integer}, ε = size(basins, 1)÷10) -> Sb, Sbb
 
 Return the basin entropy [Daza2016](@cite) `Sb` and basin boundary entropy `Sbb`
 of the given `basins` of attraction by considering `ε`-sized boxes along each dimension.
+
+The second function signature exists for backwards compatibility. 
 
 ## Description
 
@@ -44,6 +47,11 @@ function basin_entropy(basins::AbstractArray{<:Integer, D}, ε::Integer = size(b
     return basin_entropy(basins, es)
 end
 
+function basin_entropy(BoA::ArrayBasinsOfAttraction{<:Integer}, ε::Integer = size(BoA.basins, 1)÷10) 
+    es = ntuple(i -> ε, Val(ndims(BoA.basins)))
+    return basin_entropy(BoA,es) 
+end
+
 function basin_entropy(basins::AbstractArray{<:Integer, D}, es::NTuple{D, <: Integer}) where {D}
     if size(basins) .% es ≠ ntuple(i -> 0, D)
         throw(ArgumentError("The basins are not fully divisible by the sizes `ε`"))
@@ -65,6 +73,8 @@ function basin_entropy(basins::AbstractArray{<:Integer, D}, es::NTuple{D, <: Int
     return Sb/length(box_iterator), Sb/Nb
 end
 
+basin_entropy(BoA::ArrayBasinsOfAttraction{<:Integer}, es::NTuple{D, <: Integer}) where {D} = basin_entropy(BoA.basins, es)
+
 function _box_entropy(box_values, unique_vals = unique(box_values))
     h = 0.0
     for v in unique_vals
@@ -77,11 +87,14 @@ end
 
 
 """
+    basin_entropy(BoA::ArrayBasinsOfAttraction, ε = 20, Ntotal = 1000) -> test_res, Sbb
     basins_fractal_test(basins; ε = 20, Ntotal = 1000) -> test_res, Sbb
 
 Perform an automated test to decide if the boundary of the basins has fractal structures
 based on the method of Puy et al. [Puy2021](@cite).
 Return `test_res` (`:fractal` or `:smooth`) and the mean basin boundary entropy.
+
+The second function signature exists for backwards compatibility. 
 
 ## Keyword arguments
 
@@ -161,14 +174,19 @@ function basins_fractal_test(basins; ε = 20, Ntotal = 1000)
     return tst_res, Ŝbb
 end
 
+basins_fractal_test(BoA::ArrayBasinsOfAttraction{<:Integer}, ε = 20, Ntotal = 1000) = basins_fractal_test(BoA.basins, ε, Ntotal)
+
 # as suggested in https://github.com/JuliaStats/StatsBase.jl/issues/398#issuecomment-417875619
 linreg(x, y) = hcat(fill!(similar(x), 1), x) \ y
 
 """
+    basins_fractal_dimension(BoA::ArrayBasinsOfAttraction; kwargs...) -> V_ε, N_ε, d
     basins_fractal_dimension(basins; kwargs...) -> V_ε, N_ε, d
 
 Estimate the fractal dimension `d` of the boundary between basins of attraction using
 a box-counting algorithm for the boxes that contain at least two different basin IDs.
+
+The second function signature exists for backwards compatibility. 
 
 ## Keyword arguments
 
@@ -225,7 +243,9 @@ function basins_fractal_dimension(basins::AbstractArray; range_ε = 3:maximum(si
     return V_ε, N_ε, d
 end
 
+basins_fractal_dimension(BoA::ArrayBasinsOfAttraction; range_ε = 3:maximum(size(BoA.basins))÷20) = basins_fractal_dimension(BoA.basins; range_ε=range_ε)
 """
+    uncertainty_exponent(BoA::ArrayBasinsOfAttraction; kwargs...) -> ε, N_ε, α
     uncertainty_exponent(basins; kwargs...) -> ε, N_ε, α
 
 Estimate the uncertainty exponent[Grebogi1983](@cite) of the basins of attraction. This exponent
@@ -238,6 +258,8 @@ that contain at least two initial conditions that lead to different attractors.
 The output `α` is the estimation of the uncertainty exponent using the box-counting
 dimension of the boundary by fitting a line in the `log.(N_ε)` vs `log.(1/ε)` curve.
 However it is recommended to analyze the curve directly for more accuracy.
+
+The second function signature exists for backwards compatibility. 
 
 ## Keyword arguments
 * `range_ε = 2:maximum(size(basins))÷20` is the range of sizes of the ball to
@@ -264,3 +286,5 @@ function uncertainty_exponent(basins::AbstractArray; range_ε = 2:maximum(size(b
     V_ε, N_ε, d = basins_fractal_dimension(basins; range_ε)
     return V_ε, N_ε, length(size(basins)) - d
 end
+
+uncertainty_exponent(BoA::ArrayBasinsOfAttraction; range_ε = 2:maximum(size(BoA.basins))÷20) = uncertainty_exponent(BoA.basins, range_ε = range_ε)